#102 Move folders, add login not specified message
