+++ services/frontend/src/components/AppRouter.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import {Navigate, Route, Routes} from "react-router-dom";
-import WorklogDebts from "../pages/WorklogDebts";
-import Employees from "../pages/Employees";
+import WorklogDebts from "./pages/worklogdebt/WorklogDebts";
+import Employees from "./pages/employee/Employees";
 import NotFound from "./error/NotFound";
-import NotificationUsers from "../pages/NotificationUsers";
+import NotificationUsers from "./pages/notification-users/NotificationUsers";
 
 const AppRouter = () => {
     return (
