+++ services/frontend/src/components/pages/employee/AddEmployeeFormContent.tsx
@@ -1,8 +1,8 @@
 import React, {Dispatch, SetStateAction} from "react";
-import {Region} from "../models/region/Region";
-import {CreateEmployeeData} from "../models/employee/CreateEmployeeData";
-import {EmployeeFormValidator} from "../validation/EmployeeFormValidator";
-import {EmployeeFormErrors} from "../validation/EmployeeFormErrors";
+import {Region} from "../../../models/region/Region";
+import {CreateEmployeeData} from "../../../models/employee/CreateEmployeeData";
+import {EmployeeFormValidator} from "../../../validation/EmployeeFormValidator";
+import {EmployeeFormErrors} from "../../../validation/EmployeeFormErrors";
 
 type AddEmployeeContentProps = {
     formData: CreateEmployeeData;
