+++ services/frontend/src/components/pages/employee/EditEmployeeModal.tsx
@@ -1,14 +1,14 @@
 import React, {useState} from 'react';
-import CustomModal from "./CustomModal";
-import {Employee} from "../../models/employee/Employee";
-import {Region} from "../../models/region/Region";
-import {useRequest} from "../../hooks/useRequest";
-import EmployeeService from "../../services/EmployeeService";
-import EditEmployeeFormContent from "../EditEmployeeFormContent";
-import FormLoader from "../loader/FormLoader";
-import SuccessToast from "../SuccessToast";
-import {UpdateEmployeeData} from "../../models/employee/UpdateEmployeeData";
-import {EmployeeFormErrors} from "../../validation/EmployeeFormErrors";
+import CustomModal from "../../modal/CustomModal";
+import {Employee} from "../../../models/employee/Employee";
+import {Region} from "../../../models/region/Region";
+import {useRequest} from "../../../hooks/useRequest";
+import EmployeeService from "../../../api/EmployeeService";
+import EditEmployeeFormContent from "./EditEmployeeFormContent";
+import FormLoader from "../../loader/FormLoader";
+import SuccessToast from "../../SuccessToast";
+import {UpdateEmployeeData} from "../../../models/employee/UpdateEmployeeData";
+import {EmployeeFormErrors} from "../../../validation/EmployeeFormErrors";
 
 type EditEmployeeModalProps = {
     employee: Employee;
