+++ services/frontend/src/components/pages/worklogdebt/WorklogDebts.tsx
@@ -1,15 +1,15 @@
 import React, {useEffect, useState} from 'react';
-import '../styles/content.scss';
-import '../styles/worklogDebts.scss';
-import WorklogDebtsList from "../components/WorklogDebtsList";
-import {useRequest} from "../hooks/useRequest";
-import WorklogDebtsService from "../services/WorklogDebtsService";
-import {EmployeeWorklogDebts} from "../models/worklogdebt/EmployeeWorklogDebts";
-import Loader from "../components/loader/Loader";
-import CustomButton from "../components/button/CustomButton";
-import DebtsNotificationService from "../services/DebtsNotificationService";
-import {NotificationResponse} from "../models/notification/NotificationResponse";
-import {NotificationStatus} from "../models/notification/NotificationStatus";
+import '../../../styles/content.scss';
+import '../../../styles/worklogDebts.scss';
+import WorklogDebtsList from "./WorklogDebtsList";
+import {useRequest} from "../../../hooks/useRequest";
+import WorklogDebtsService from "../../../api/WorklogDebtsService";
+import {EmployeeWorklogDebts} from "../../../models/worklogdebt/EmployeeWorklogDebts";
+import Loader from "../../loader/Loader";
+import CustomButton from "../../button/CustomButton";
+import DebtsNotificationService from "../../../api/DebtsNotificationService";
+import {NotificationResponse} from "../../../models/notification/NotificationResponse";
+import {NotificationStatus} from "../../../models/notification/NotificationStatus";
 
 export enum NotificationLoadingStatus {
     Inactive,
